@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "todo",
   "main": "index.js",
-  "proxy": "http://localhost:3000",
+  "proxy": "http://127.0.0.1:3000",
   "scripts": {
     "start": "nodemon server/server.js",
     "lint": "./node_modules/.bin/eslint",
@@ -23,20 +23,26 @@
   },
   "homepage": "https://github.com/samuelberston/MVP#readme",
   "dependencies": {
+    "@auth0/auth0-react": "^2.0.0",
     "@fortawesome/fontawesome-free": "^5.15.4",
+    "auth0-lock": "^12.0.2",
     "axios": "^0.21.1",
     "buffer": "^6.0.3",
+    "cors": "^2.8.5",
     "crypto-browserify": "^3.12.0",
     "css-loader": "^6.7.1",
     "dotenv": "^16.0.3",
     "express": "^4.17.1",
-    "nock": "^13.3.0",
+    "history": "^5.3.0",
+    "jest": "^29.3.1",
     "node-cron": "^3.0.0",
     "nodemon": "^2.0.20",
     "process": "^0.11.10",
-    "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "prop-types": "^15.7.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "^6.8.1",
+    "react-router-dom": "^6.8.1",
     "react-select": "^5.6.1",
     "sass": "^1.32.8",
     "sinon": "^15.0.1",
@@ -55,6 +61,7 @@
     "babel-cli": "^6.26.0",
     "babel-jest": "^29.4.0",
     "babel-loader": "^8.2.2",
+    "dotenv-webpack": "^8.0.1",
     "enzyme": "^3.11.0",
     "eslint": "^7.23.0",
     "eslint-config-airbnb": "^18.2.1",
